moves default and custom receipt out of process wrapper and into separate routes
